requirements.txt
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.11
Django==1.10.6
django-bower==5.2.0
enum34==1.1.6
gunicorn==19.7.0
ipython==5.1.0
ipython-genutils==0.1.0
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.13
psycopg2==2.7.1
ptyprocess==0.5.1
Pygments==2.2.0
python-nmap==0.6.1
scandir==1.5
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
virtualenv==15.0.3
wcwidth==0.1.7
